--- packages/builtin-agent-onboarding/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@lobechat/builtin-agent-onboarding",
-  "version": "1.0.0",
-  "private": true,
-  "exports": {
-    ".": "./src/index.ts",
-    "./client": "./src/client/index.ts"
-  },
-  "main": "./src/index.ts",
-  "devDependencies": {
-    "@lobechat/types": "workspace:*"
-  },
-  "peerDependencies": {
-    "@lobehub/ui": "^5",
-    "antd": "^6",
-    "react": "*"
-  }
-}
