--- packages/builtin-agent-onboarding/src/client/index.ts
@@ -1,7 +1,0 @@
-export type {
-  DefaultModelConfig,
-  QuestionRendererProps,
-  QuestionRendererRenderEmojiPickerProps,
-  QuestionRendererRenderModelSelectProps,
-} from './QuestionRenderer';
-export { default as QuestionRenderer } from './QuestionRenderer';
