--- packages/builtin-agent-onboarding/src/index.ts
@@ -1,2 +1,0 @@
-export { createSystemRole } from './systemRole';
-export { toolSystemPrompt } from './toolSystemRole';
