+++ packages/builtin-agents/package.json
@@ -4,7 +4,6 @@
   "private": true,
   "main": "./src/index.ts",
   "dependencies": {
-    "@lobechat/builtin-agent-onboarding": "workspace:*",
     "@lobechat/builtin-tool-agent-builder": "workspace:*",
     "@lobechat/builtin-tool-agent-documents": "workspace:*",
     "@lobechat/builtin-tool-agent-management": "workspace:*",
@@ -13,7 +12,6 @@
     "@lobechat/builtin-tool-gtd": "workspace:*",
     "@lobechat/builtin-tool-notebook": "workspace:*",
     "@lobechat/builtin-tool-user-interaction": "workspace:*",
-    "@lobechat/builtin-tool-web-onboarding": "workspace:*",
     "@lobechat/business-const": "workspace:*",
     "@lobechat/const": "workspace:*"
   },
