+++ packages/builtin-agents/src/agents/web-onboarding/index.ts
@@ -1,13 +1,15 @@
-import { createSystemRole } from '@lobechat/builtin-agent-onboarding';
 import { AgentManagementIdentifier } from '@lobechat/builtin-tool-agent-management';
 import { GroupAgentBuilderIdentifier } from '@lobechat/builtin-tool-group-agent-builder';
 import { UserInteractionIdentifier } from '@lobechat/builtin-tool-user-interaction';
-import { WebOnboardingIdentifier } from '@lobechat/builtin-tool-web-onboarding';
 import { DEFAULT_PROVIDER } from '@lobechat/business-const';
 import { DEFAULT_ONBOARDING_MODEL } from '@lobechat/const';
 
 import type { BuiltinAgentDefinition } from '../../types';
 import { BUILTIN_AGENT_SLUGS } from '../../types';
+import { createSystemRole } from './systemRole';
+
+/** Must match `WebOnboardingIdentifier` in `packages/builtin-tool-web-onboarding/src/types.ts`. */
+const WebOnboardingIdentifier = 'lobe-web-onboarding';
 
 export const WEB_ONBOARDING: BuiltinAgentDefinition = {
   avatar: '/avatars/lobe-ai.png',
