+++ packages/builtin-tool-web-onboarding/package.json
@@ -9,7 +9,6 @@
   },
   "main": "./src/index.ts",
   "dependencies": {
-    "@lobechat/builtin-agent-onboarding": "workspace:*",
     "@lobechat/prompts": "workspace:*"
   },
   "devDependencies": {
