+++ src/routes/onboarding/_layout/index.tsx
@@ -1,9 +1,10 @@
 'use client';
 
+import { AGENT_ONBOARDING_ENABLED } from '@lobechat/business-const';
 import { Center, Flexbox, FluentEmoji, Text } from '@lobehub/ui';
 import { Divider, Popconfirm } from 'antd';
 import { cx, useTheme } from 'antd-style';
-import { type FC, type MouseEvent, type PropsWithChildren, useCallback } from 'react';
+import { type FC, type MouseEvent, type PropsWithChildren, useCallback, useMemo } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { useLocation, useNavigate } from 'react-router-dom';
 
@@ -23,6 +24,12 @@ const OnBoardingContainer: FC<PropsWithChildren> = ({ children }) => {
   const navigate = useNavigate();
   const finishOnboarding = useUserStore((s) => s.finishOnboarding);
   const isAgentOnboarding = pathname.startsWith('/onboarding/agent');
+
+  const showModeSwitchAndSkipFooter = useMemo(() => {
+    if (!isAgentOnboarding) return true;
+
+    return AGENT_ONBOARDING_ENABLED;
+  }, [isAgentOnboarding]);
 
   const handleConfirmSkip = useCallback(() => {
     finishOnboarding();
@@ -65,56 +72,60 @@ const OnBoardingContainer: FC<PropsWithChildren> = ({ children }) => {
         <Center height={'100%'} width={'100%'}>
           {children}
         </Center>
-        <Center paddingBlock={'0 8px'} paddingInline={16}>
-          <Text fontSize={12} type={'secondary'}>
-            <Trans
-              i18nKey={'agent.layout.switchMessage'}
-              ns={'onboarding'}
-              components={{
-                modeLink: (
-                  <a
-                    href={isAgentOnboarding ? '/onboarding/classic' : '/onboarding/agent'}
-                    onClick={swichMode}
-                  />
-                ),
-                modeText: <Text as={'span'} />,
-                skipLink: (
-                  <Popconfirm
-                    arrow={false}
-                    cancelButtonProps={{ type: 'text' }}
-                    cancelText={t('cancel', { ns: 'common' })}
-                    okText={t('agent.layout.skipConfirm.ok', { ns: 'onboarding' })}
-                    style={{ cursor: 'pointer' }}
-                    description={
-                      <Text fontSize={13} style={{ marginBottom: 8 }} type={'secondary'}>
-                        {t('agent.layout.skipConfirm.content', { ns: 'onboarding' })}
-                      </Text>
-                    }
-                    icon={
-                      <FluentEmoji
-                        emoji={'😗'}
-                        size={24}
-                        style={{ marginRight: 8 }}
-                        type={'anim'}
-                      />
-                    }
-                    title={
-                      <Text fontSize={15}>{t('agent.completionTitle', { ns: 'onboarding' })}</Text>
-                    }
-                    onConfirm={handleConfirmSkip}
-                  />
-                ),
-                skipText: <Text as={'span'} style={{ cursor: 'pointer' }} />,
-              }}
-              values={{
-                mode: isAgentOnboarding
-                  ? t('agent.layout.mode.classic')
-                  : t('agent.layout.mode.agent'),
-                skip: t('agent.layout.skip'),
-              }}
-            />
-          </Text>
-        </Center>
+        {showModeSwitchAndSkipFooter && (
+          <Center paddingBlock={'0 8px'} paddingInline={16}>
+            <Text fontSize={12} type={'secondary'}>
+              <Trans
+                i18nKey={'agent.layout.switchMessage'}
+                ns={'onboarding'}
+                components={{
+                  modeLink: (
+                    <a
+                      href={isAgentOnboarding ? '/onboarding/classic' : '/onboarding/agent'}
+                      onClick={swichMode}
+                    />
+                  ),
+                  modeText: <Text as={'span'} />,
+                  skipLink: (
+                    <Popconfirm
+                      arrow={false}
+                      cancelButtonProps={{ type: 'text' }}
+                      cancelText={t('cancel', { ns: 'common' })}
+                      okText={t('agent.layout.skipConfirm.ok', { ns: 'onboarding' })}
+                      style={{ cursor: 'pointer' }}
+                      description={
+                        <Text fontSize={13} style={{ marginBottom: 8 }} type={'secondary'}>
+                          {t('agent.layout.skipConfirm.content', { ns: 'onboarding' })}
+                        </Text>
+                      }
+                      icon={
+                        <FluentEmoji
+                          emoji={'😗'}
+                          size={24}
+                          style={{ marginRight: 8 }}
+                          type={'anim'}
+                        />
+                      }
+                      title={
+                        <Text fontSize={15}>
+                          {t('agent.completionTitle', { ns: 'onboarding' })}
+                        </Text>
+                      }
+                      onConfirm={handleConfirmSkip}
+                    />
+                  ),
+                  skipText: <Text as={'span'} style={{ cursor: 'pointer' }} />,
+                }}
+                values={{
+                  mode: isAgentOnboarding
+                    ? t('agent.layout.mode.classic')
+                    : t('agent.layout.mode.agent'),
+                  skip: t('agent.layout.skip'),
+                }}
+              />
+            </Text>
+          </Center>
+        )}
       </Flexbox>
     </Flexbox>
   );
